==> analysis/03_analysis.Rmd <==
---
title: "Analysis"
author: "COMBINE Australia"
date: '`r Sys.Date()`'
output:
  html_document:
    number_sections: yes
    theme: yeti
    toc: yes
    toc_float: yes
    df_print: paged
---

Code version: `r system("git log -1 --format=oneline | cut -d' ' -f1", intern = TRUE)`

```{r knitr, include = FALSE}
DOCNAME = "analysis"
knitr::opts_chunk$set(autodep        = TRUE,
                      cache          = FALSE,
                      cache.path     = paste0("cache/", DOCNAME, "/"),
                      cache.comments = TRUE,
                      echo           = FALSE,
                      error          = FALSE,
                      fig.align      = "center",
                      fig.path       = paste0("figures/", DOCNAME, "/"),
                      fig.width      = 10,
                      fig.height     = 8,
                      message        = FALSE,
                      warning        = FALSE)
```

```{r libaries, cache = FALSE}
library("here")
library("tidyverse")
```

```{r source, cache = FALSE}

```

Introduction
============

In this document we are going to do a more thorough analysis of the survey
responses, splitting them up in different ways etc. This will help us decide
what to include in the final report.

```{r load}
col_types <- cols(
    Timestamp                     = col_datetime(format = ""),
    Gender                        = col_character(),
    Age                           = col_integer(),
    IndigenousStatus              = col_character(),
    Location                      = col_character(),
    CourseLevel                   = col_character(),
    CourseYear                    = col_character(),
    Residency                     = col_character(),
    TimeCommitment                = col_character(),
    Background                    = col_character(),
    BackgroundGrouped             = col_character(),
    Competencies_Genetics         = col_integer(),
    Competencies_Biochemistry     = col_integer(),
    Competencies_Probability      = col_integer(),
    Competencies_Testing          = col_integer(),
    Competencies_LinearAlgebra    = col_integer(),
    Competencies_Algorithms       = col_integer(),
    Competencies_Programming      = col_integer(),
    Competencies_SoftwareDev      = col_integer(),
    Finances                      = col_character(),
    Finances_Government           = col_logical(),
    Finances_University           = col_logical(),
    Finances_TopUp                = col_logical(),
    Finances_PartTime             = col_logical(),
    Finances_FullTime             = col_logical(),
    Finances_Parental             = col_logical(),
    Finances_Savings              = col_logical(),
    Finances_HasOther             = col_logical(),
    Finances_Other                = col_character(),
    CourseworkComponent           = col_character(),
    CourseworkCurrent             = col_character(),
    Classes                       = col_double(),
    Classes_Computational         = col_integer(),
    Classes_Statistical           = col_integer(),
    Classes_Biological             = col_integer(),
    CourseworkTime                = col_double(),
    Assignments                   = col_integer(),
    Exams                         = col_integer(),
    Cohort                        = col_character(),
    Support                       = col_character(),
    CourseworkDetails             = col_character(),
    ResearchComponent             = col_character(),
    ResearchCurrent               = col_character(),
    ResearchPercentage            = col_integer(),
    ResearchTime                  = col_integer(),
    ResearchFocus                 = col_character(),
    Focus_GenomicVariants         = col_logical(),
    Focus_GeneExpression          = col_logical(),
    Focus_Metagenomics            = col_logical(),
    Focus_Proteomics              = col_logical(),
    Focus_DataVis                 = col_logical(),
    Focus_Methods                 = col_logical(),
    Focus_Cancer                  = col_logical(),
    Focus_Analysis                = col_logical(),
    Focus_CellBiology             = col_logical(),
    Focus_PlantBiology            = col_logical(),
    Focus_RareDisease             = col_logical(),
    Focus_Metabolomics            = col_logical(),
    Focus_MicrobialBiology        = col_logical(),
    Focus_HumanBiology            = col_logical(),
    Focus_ModelOrganisims         = col_logical(),
    Focus_NonModelOrganisims      = col_logical(),
    Focus_HasOther                = col_logical(),
    Focus_Other                   = col_character(),
    Institute                     = col_character(),
    Computational                 = col_integer(),
    Technologies                  = col_character(),
    Technologies_RNASeq           = col_logical(),
    Technologies_WholeGenome      = col_logical(),
    Technologies_Exome            = col_logical(),
    Technologies_ShortRead        = col_logical(),
    Technologies_LongRead         = col_logical(),
    Technologies_MassSpec         = col_logical(),
    Technologies_Array            = col_logical(),
    Technologies_FlowCytometry    = col_logical(),
    Technologies_HasOther         = col_logical(),
    Technologes_Other             = col_character(),
    ProgrammingLanguages          = col_character(),
    Programming_R                 = col_logical(),
    Programming_Python            = col_logical(),
    Programming_Shell             = col_logical(),
    Programming_Git               = col_logical(),
    Programming_Perl              = col_logical(),
    Programming_CCpp              = col_logical(),
    Programming_MATLAB            = col_logical(),
    Programming_HasOther          = col_logical(),
    Programming_Other             = col_character(),
    Collaboration                 = col_character(),
    Collaboration_YourGroup       = col_logical(),
    Collaboration_OtherGroups     = col_logical(),
    Collaboration_ExternalGroups  = col_logical(),
    Collaboration_OtherStudents   = col_logical(),
    ResearchGroup                 = col_character(),
    OtherStudents                 = col_character(),
    ResearchDetails               = col_character(),
    SupervisorNumber              = col_integer(),
    PrimaryPercentage             = col_double(),
    PrimaryGender                 = col_character(),
    SecondaryGender               = col_character(),
    FormalMeetings                = col_character(),
    InformalMeetings              = col_character(),
    Supervisor_Communication      = col_integer(),
    Supervisor_Inclusiveness      = col_integer(),
    Supervisor_Empathy            = col_integer(),
    Supervisor_SubjectMatter      = col_integer(),
    Supervisor_Availability       = col_integer(),
    Supervisor_Project            = col_integer(),
    ExternalActivities            = col_character(),
    FacilityAccess                = col_character(),
    Facilities_Desk               = col_logical(),
    Facilities_Computer           = col_logical(),
    Facilities_Server             = col_logical(),
    Facilities_Lab                = col_logical(),
    Facilities_MeetingRooms       = col_logical(),
    Facilities_BreakRooms         = col_logical(),
    FinancialSupport              = col_character(),
    FinancialSupport_TravelApp    = col_logical(),
    FinancialSupport_TravelNoApp  = col_logical(),
    FinancialSupport_TopUp        = col_logical(),
    FinancialSupport_HasOther     = col_logical(),
    FinancialSupport_Other        = col_character(),
    OtherSupport                  = col_character(),
    OtherSupport_Group            = col_logical(),
    OtherSupport_Institute        = col_logical(),
    OtherSupport_Students         = col_logical(),
    OtherSupport_People           = col_logical(),
    OtherSupport_None             = col_logical(),
    SupervisorDetails             = col_character(),
    FamiliarityCOMBINE            = col_character(),
    COMBINE_Survey                = col_logical(),
    COMBINE_SocialMedia           = col_logical(),
    COMBINE_Newsletter            = col_logical(),
    COMBINE_Workshop              = col_logical(),
    COMBINE_Seminar               = col_logical(),
    COMBINE_Symposium             = col_logical(),
    FamiliarityABACBS             = col_character(),
    ABACBS_Member                 = col_logical(),
    ABACBS_SocialMedia            = col_logical(),
    ABACBS_Newsletter             = col_logical(),
    ABABCS_Conference             = col_logical(),
    ABACBS_Event                  = col_logical(),
    ABACBS_None                   = col_logical(),
    MembershipABACBS              = col_character(),
    Membership_National           = col_logical(),
    Membership_EventDiscount      = col_logical(),
    Membership_ConferenceDiscount = col_logical(),
    Membership_Travel             = col_logical(),
    Membership_Prizes             = col_logical(),
    Membership_CV                 = col_logical(),
    Membership_Direction          = col_logical(),
    Membership_NoBenefits         = col_logical(),
    EventsCOMBINE                 = col_character(),
    Events_Programming            = col_logical(),
    Events_Bioinformatics         = col_logical(),
    Events_Skills                 = col_logical(),
    Events_Statistics             = col_logical(),
    Events_Lab                    = col_logical(),
    Events_OneOff                 = col_logical(),
    Events_InvitedSeminar         = col_logical(),
    Events_StudentSeminar         = col_logical(),
    Events_Social                 = col_logical(),
    Events_Symposium              = col_logical(),
    Events_NotInterested          = col_logical(),
    Events_HasOther               = col_logical(),
    Events_Other                  = col_character(),
    WhatCanWeDo                   = col_character()
)

responses <- read_tsv(here("data/responses_clean.tsv"),
                      col_types = col_types)
```


Demographics
============

Gender
------

```{r gender}
ggplot(responses, aes(x = Gender, y = ..prop.., group = 1)) + 
  geom_bar() + 
  scale_y_continuous(labels = scales::percent_format())
```

Age
---

```{r age}
ggplot(responses, aes(x = 1, y = Age)) +
    geom_boxplot()
```

### By gender

```{r age-gender}
ggplot(responses, aes(x = Gender, y = Age, colour = Gender)) +
    geom_boxplot()
```

### By course

```{r age-course}
ggplot(responses, aes(x = CourseLevel, y = Age)) +
    geom_boxplot() +
    coord_flip()
```

### By location

```{r age-location}
ggplot(responses, aes(x = Location, y = Age)) +
    geom_boxplot() +
    coord_flip()
```

### By residency

```{r age-residency}
ggplot(responses, aes(x = Residency, y = Age)) +
    geom_boxplot()
```

Location
--------

```{r location}
ggplot(responses, aes(x = Location)) +
    geom_bar()
```

### By gender

```{r location-gender}
ggplot(responses, aes(x = Location, fill = Gender)) +
    geom_bar()
```

Course level
------------

```{r course-level}
ggplot(responses, aes(x = CourseLevel, y = ..prop.., group = 1)) + 
  geom_bar() + 
  scale_y_continuous(labels = scales::percent_format())
```

### By gender

```{r course-level-gender}
ggplot(responses,
       aes(x = CourseLevel, fill = Gender)) + 
  geom_bar()
```

### By residency

```{r course-level-residency}
ggplot(responses, aes(x = CourseLevel)) +
    geom_bar() +
    facet_wrap(~ Residency)
```

### By time commitment

```{r course-level-time}
ggplot(responses, aes(x = CourseLevel, fill = TimeCommitment)) +
    geom_bar()
```

Course year
-----------

```{r course-year}
ggplot(responses, aes(x = CourseYear)) +
    geom_bar()
```

### By gender

```{r course-year-gender}
ggplot(responses, aes(x = CourseYear, fill = Gender)) +
    geom_bar()
```

### By course

```{r course-year-course}
ggplot(responses, aes(x = CourseYear)) +
    geom_bar() +
    facet_wrap(~ CourseLevel, ncol = 1)
```

Residency
---------

```{r residency}
ggplot(responses, aes(x = Residency, y = ..prop.., group = 1)) + 
  geom_bar() + 
  scale_y_continuous(labels = scales::percent_format())
```

### By gender

```{r residency-gender}
ggplot(responses, aes(x = Residency, fill = Gender)) +
    geom_bar()
```

Time commitment
---------------

```{r time}
ggplot(responses, aes(x = TimeCommitment)) +
    geom_bar()
```

### By gender

```{r time-gender}
ggplot(responses, aes(x = TimeCommitment, fill = Gender)) +
    geom_bar()
```

Background
----------

```{r background}
ggplot(responses, aes(x = BackgroundGrouped)) +
    geom_bar()
```

### By gender

```{r background-gender}
ggplot(responses, aes(x = BackgroundGrouped, fill = Gender)) +
    geom_bar()
```

Competencies
------------

```{r competencies}
responses %>%
    select(starts_with("Competencies")) %>%
    gather(key = "Competency", value = "Rating") %>%
    mutate(Competency = str_remove(Competency, "Competencies_")) %>%
    group_by(Competency, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Competency, y = Prop, fill = Rating)) +
    geom_col() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))
```

### By gender

```{r competencies-gender}
responses %>%
    select(Gender, starts_with("Competencies")) %>%
    gather(key = "Competency", value = "Rating", -Gender) %>%
    mutate(Competency = str_remove(Competency, "Competencies_")) %>%
    group_by(Gender, Competency, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Competency, y = Prop, fill = Rating)) +
    geom_col() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
    facet_wrap(~ Gender, ncol = 1)
```

### By background

```{r competencies-background}
responses %>%
    select(BackgroundGrouped, starts_with("Competencies")) %>%
    gather(key = "Competency", value = "Rating", -BackgroundGrouped) %>%
    mutate(Competency = str_remove(Competency, "Competencies_")) %>%
    group_by(BackgroundGrouped, Competency, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Competency, y = Prop, fill = Rating)) +
    geom_col() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
    facet_wrap(~ BackgroundGrouped, ncol = 1)
```

Finances
--------

```{r finances}
responses %>%
    select(starts_with("Finances_")) %>%
    gather(key = "Finance", value = "Has") %>%
    mutate(Finance = str_remove(Finance, "Finances_")) %>%
    filter(Has == TRUE) %>%
    ggplot(aes(x = Finance)) +
    geom_bar()
```

### By gender

```{r finances-gender}
responses %>%
    select(starts_with("Finances_"), Gender) %>%
    gather(key = "Finance", value = "Has", -Gender) %>%
    mutate(Finance = str_remove(Finance, "Finances_")) %>%
    filter(Has == TRUE) %>%
    ggplot(aes(x = Finance)) +
    geom_bar() +
    facet_wrap(~ Gender, ncol = 1)
```

### By course

```{r finances-course}
responses %>%
    select(starts_with("Finances_"), CourseLevel) %>%
    filter(CourseLevel %in% c("PhD", "Masters (Coursework)")) %>%
    gather(key = "Finance", value = "Has", -CourseLevel) %>%
    mutate(Finance = str_remove(Finance, "Finances_")) %>%
    filter(Has == TRUE) %>%
    group_by(CourseLevel) %>%
    mutate(CourseCount = n()) %>%
    group_by(CourseLevel, Finance, CourseCount) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / CourseCount) %>%
    ggplot(aes(x = Finance, y = Prop)) +
    geom_col() +
    facet_wrap(~ CourseLevel, ncol = 1)
```

Coursework
==========

```{r coursework}
coursework <- responses %>%
    filter(CourseworkCurrent == "Yes")
```

Classes
-------

```{r classes}
ggplot(coursework, aes(x = factor(Classes), y = CourseworkTime,
                       colour = CourseLevel)) +
    geom_jitter(height = 0, width = 0.1)
```

Assignments
-----------

```{r assignments}
ggplot(coursework, aes(x = factor(Classes), y = Assignments,
                       colour = CourseLevel)) +
    geom_jitter(height = 0, width = 0.1)
```

Exams
-----

```{r exams}
ggplot(coursework, aes(x = factor(Classes), y = Exams,
                       colour = CourseLevel)) +
    geom_jitter(height = 0, width = 0.1)
```

Class types
-----------

```{r class-types}
coursework %>%
    select(Classes_Computational, Classes_Statistical, Classes_Biological) %>%
    filter(!is.na(Classes_Computational)) %>%
    arrange(Classes_Statistical, Classes_Computational, Classes_Biological) %>%
    mutate(Person = row_number()) %>%
    gather(key = Type, value = Percent, -Person) %>%
    mutate(Type = str_remove(Type, "Classes_")) %>%
    ggplot(aes(x = factor(Person), y = Percent, fill = Type)) +
    geom_col()
```

Support
-------

```{r coursework-support}
coursework %>%
    select(Cohort, Support) %>%
    gather(key = Type, value = IsTrue) %>%
    filter(IsTrue == "Yes") %>%
    ggplot(aes(x = Type)) +
    geom_bar()
```

### By gender

```{r coursework-support-gender}
coursework %>%
    select(Cohort, Support, Gender) %>%
    gather(key = Type, value = IsTrue, -Gender) %>%
    filter(IsTrue == "Yes") %>%
    ggplot(aes(x = Type, fill = Gender)) +
    geom_bar() +
    facet_wrap(~ Gender, ncol = 1)
```

### By residency

```{r coursework-support-residency}
coursework %>%
    select(Cohort, Support, Residency) %>%
    gather(key = Type, value = IsTrue, -Residency) %>%
    filter(IsTrue == "Yes") %>%
    ggplot(aes(x = Type, fill = Residency)) +
    geom_bar() +
    facet_wrap(~ Residency, ncol = 1)
```

Research
========

```{r research}
research <- responses %>%
    filter(ResearchCurrent == "Yes")
```

Percentage
----------

```{r percentage}
ggplot(research, aes(x = ResearchPercentage)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

Time
----

```{r researchtime}
ggplot(research, aes(x = ResearchTime)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

Focus
-----

```{r focus}
research %>%
    select(starts_with("Focus_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Focus", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Focus = str_remove(Focus, "Focus_")) %>%
    group_by(Focus, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Focus = factor(Focus, levels = Focus),
           Focus = forcats::fct_relevel(Focus, "HasOther", after = 0)) %>%
    ggplot(aes(x = Focus, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By gender

```{r focus-gender}
research %>%
    select(starts_with("Focus_"), Gender) %>%
    group_by(Gender) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Focus", value = "Has", -GroupSize, -Gender) %>%
    filter(Has == TRUE) %>%
    mutate(Focus = str_remove(Focus, "Focus_")) %>%
    group_by(Focus, GroupSize, Gender) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Focus = factor(Focus, levels = unique(Focus)),
           Focus = forcats::fct_relevel(Focus, "HasOther", after = 0)) %>%
    ggplot(aes(x = Focus, y = Prop, fill = Gender)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Gender, nrow = 1)
```

### By residency

```{r focus-residency}
research %>%
    select(starts_with("Focus_"), Residency) %>%
    group_by(Residency) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Focus", value = "Has", -GroupSize, -Residency) %>%
    filter(Has == TRUE) %>%
    mutate(Focus = str_remove(Focus, "Focus_")) %>%
    group_by(Focus, GroupSize, Residency) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Focus = factor(Focus, levels = unique(Focus)),
           Focus = forcats::fct_relevel(Focus, "HasOther", after = 0)) %>%
    ggplot(aes(x = Focus, y = Prop, fill = Residency)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Residency, nrow = 1)
```

### By background

```{r focus-background}
research %>%
    select(starts_with("Focus_"), BackgroundGrouped) %>%
    group_by(BackgroundGrouped) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Focus", value = "Has", -GroupSize, -BackgroundGrouped) %>%
    filter(Has == TRUE) %>%
    mutate(Focus = str_remove(Focus, "Focus_")) %>%
    group_by(Focus, GroupSize, BackgroundGrouped) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Focus = factor(Focus, levels = unique(Focus)),
           Focus = forcats::fct_relevel(Focus, "HasOther", after = 0)) %>%
    ggplot(aes(x = Focus, y = Prop, fill = BackgroundGrouped)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ BackgroundGrouped, nrow = 1) +
    theme(legend.position = "bottom")
```

Institute
---------

```{r institute}
ggplot(research, aes(x = Institute)) +
    geom_bar()
```

### By gender

```{r institute-gender}
ggplot(research, aes(x = Institute, fill = Gender)) +
    geom_bar()
```

### By residency

```{r institute-residency}
ggplot(research, aes(x = Institute, fill = Residency)) +
    geom_bar()
```

### By background

```{r institute-background}
ggplot(research, aes(x = Institute, fill = BackgroundGrouped)) +
    geom_bar(position = "dodge")
```

Computational
-------------

```{r computational}
ggplot(research, aes(x = Computational)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

### By gender

```{r computational-gender}
ggplot(research, aes(x = Computational, fill = Gender)) +
    geom_histogram(binwidth = 10, boundary = 100) +
    facet_wrap(~ Gender, ncol = 1)
```

### By background

```{r computational-background}
ggplot(research, aes(x = Computational, fill = BackgroundGrouped)) +
    geom_histogram(binwidth = 10, boundary = 100) +
    facet_wrap(~ BackgroundGrouped, ncol = 1, scales = "free_y")
```

Technologies
------------

```{r technologies}
research %>%
    select(starts_with("Technologies")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Technology", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Technology = str_remove(Technology, "Technologies_")) %>%
    group_by(Technology, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Technology = factor(Technology, levels = Technology),
           Technology = forcats::fct_relevel(Technology, "HasOther", after = 0)) %>%
    ggplot(aes(x = Technology, y = Prop)) +
    geom_col() +
    coord_flip()
```

Programming
-----------

```{r programming}
research %>%
    select(starts_with("Programming")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Language", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Language = str_remove(Language, "Programming_")) %>%
    group_by(Language, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Language = factor(Language, levels = Language),
           Language = forcats::fct_relevel(Language, "HasOther", after = 0)) %>%
    ggplot(aes(x = Language, y = Prop)) +
    geom_col() +
    coord_flip()
```

Collaboration
-------------

```{r collaboration}
research %>%
    select(starts_with("Collaboration")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Collaboration_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By gender

```{r collaboration-gender}
research %>%
    select(starts_with("Collaboration"), Gender) %>%
    group_by(Gender) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Gender) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Collaboration_")) %>%
    group_by(Type, GroupSize, Gender) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Gender)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Gender, ncol = 1)
```

### By residency

```{r collaboration-residency}
research %>%
    select(starts_with("Collaboration"), Residency) %>%
    group_by(Residency) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Residency) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Collaboration_")) %>%
    group_by(Type, GroupSize, Residency) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Residency)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Residency, ncol = 1)
```

### By background

```{r collaboration-background}
research %>%
    select(starts_with("Collaboration"), BackgroundGrouped) %>%
    group_by(BackgroundGrouped) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -BackgroundGrouped) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Collaboration_")) %>%
    group_by(Type, GroupSize, BackgroundGrouped) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = BackgroundGrouped)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ BackgroundGrouped, ncol = 1) +
    theme(legend.position = "bottom")
```

Group
-----

```{r group}
ggplot(research, aes(x = ResearchGroup)) +
    geom_bar()
```

### By gender

```{r group-gender}
ggplot(research, aes(x = ResearchGroup, fill = Gender)) +
    geom_bar()
```

### By residency

```{r group-residency}
ggplot(research, aes(x = ResearchGroup, fill = Residency)) +
    geom_bar() +
    facet_wrap(~ Residency)
```

Students
--------

```{r students}
ggplot(research, aes(x = OtherStudents)) +
    geom_bar()
```

Supervisors
===========

Number
------

```{r number}
ggplot(research, aes(x = factor(SupervisorNumber))) +
    geom_bar()
```

### By gender

```{r number-gender}
ggplot(research, aes(x = factor(SupervisorNumber), fill = Gender)) +
    geom_bar()
```

### By residency

```{r number-residency}
ggplot(research, aes(x = factor(SupervisorNumber), fill = Residency)) +
    geom_bar()
```

### By course

```{r number-course}
ggplot(research, aes(x = factor(SupervisorNumber), fill = CourseLevel)) +
    geom_bar()
```

Primary percentage
------------------

```{r primarypercentage}
ggplot(research, aes(x = PrimaryPercentage)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

### By gender

```{r primarypercentage-gender}
ggplot(research, aes(x = PrimaryPercentage, fill = Gender)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

### By primary gender

```{r primarypercentage-primarygender}
ggplot(research, aes(x = PrimaryPercentage, fill = PrimaryGender)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

### By secondary gender

```{r primarypercentage-secondarygender}
ggplot(research, aes(x = PrimaryPercentage, fill = SecondaryGender)) +
    geom_histogram(binwidth = 10, boundary = 100)
```

Primary gender
--------------

```{r primarygender}
ggplot(research, aes(x = PrimaryGender)) +
    geom_bar()
```

### By gender

```{r primarygender-gender}
ggplot(research, aes(x = PrimaryGender, fill = Gender)) +
    geom_bar()
```

### By residency

```{r primarygender-residency}
ggplot(research, aes(x = PrimaryGender, fill = Residency)) +
    geom_bar()
```

Secondary gender
----------------

```{r secondarygender}
ggplot(research, aes(x = SecondaryGender)) +
    geom_bar()
```

### By gender

```{r secondarygender-gender}
ggplot(research, aes(x = SecondaryGender, fill = Gender)) +
    geom_bar()
```

### By primary gender

```{r secondarygender-primarygender}
ggplot(research, aes(x = SecondaryGender, fill = PrimaryGender)) +
    geom_bar()
```

Both gender
-----------

```{r}
research %>%
    mutate(BothGender = paste0(PrimaryGender, SecondaryGender)) %>%
    ggplot(aes(x = BothGender, fill = Gender)) +
    geom_bar()
```

Formal meetings
---------------

```{r formalmeetings}
ggplot(research, aes(x = FormalMeetings)) +
    geom_bar()
```

### By gender

```{r formalmeetings-gender}
research %>%
    mutate(FormalMeetings = factor(FormalMeetings,
                                   levels = c("Once a week",
                                              "Once a fortnight",
                                              "Once a month",
                                              "Once every six months",
                                              "Once a year"))) %>%
ggplot(aes(x = FormalMeetings, fill = Gender)) +
    geom_bar() +
    facet_wrap(~ Gender, ncol = 1)
```

### By primary gender

```{r formalmeetings-primarygender}
research %>%
    mutate(FormalMeetings = factor(FormalMeetings,
                                   levels = c("Once a week",
                                              "Once a fortnight",
                                              "Once a month",
                                              "Once every six months",
                                              "Once a year"))) %>%
ggplot(aes(x = FormalMeetings, fill = PrimaryGender)) +
    geom_bar() +
    facet_wrap(~ PrimaryGender, ncol = 1) +
    theme(legend.position = "bottom")
```

Informal meetings
---------------

```{r informalmeetings}
research <- research %>%
    mutate(InformalMeetings = factor(InformalMeetings,
                                     levels = c("Every day",
                                                "A few times a week",
                                                "Once a week",
                                                "Once a fortnight",
                                                "Once a month",
                                                "Once every six months",
                                                "Once a year",
                                                "I never see my supervisor(s)")))

ggplot(research, aes(x = InformalMeetings)) +
    geom_bar()
```

### By gender

```{r informalmeetings-gender}
ggplot(research, aes(x = InformalMeetings, fill = Gender)) +
    geom_bar() +
    facet_wrap(~ Gender, ncol = 1)
```

### By primary gender

```{r informalmeetings-primarygender}
ggplot(research, aes(x = InformalMeetings, fill = PrimaryGender)) +
    geom_bar() +
    facet_wrap(~ PrimaryGender, ncol = 1) +
    theme(legend.position = "bottom")
```

Qualities
---------

```{r qualities}
research %>%
    select(starts_with("Supervisor_")) %>%
    gather(key = "Quality", value = "Rating") %>%
    mutate(Quality = str_remove(Quality, "Supervisor_")) %>%
    group_by(Quality, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Quality, y = Prop, fill = factor(Rating))) +
    geom_col() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))
```

### By primary gender

```{r qualities-primarugender}
research %>%
    select(PrimaryGender, starts_with("Supervisor_")) %>%
    filter(PrimaryGender %in% c("Male", "Female")) %>%
    gather(key = "Quality", value = "Rating", -PrimaryGender) %>%
    mutate(Quality = str_remove(Quality, "Supervisor_")) %>%
    group_by(PrimaryGender, Quality, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Quality, y = Prop, fill = factor(Rating))) +
    geom_col() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
    facet_wrap(~ PrimaryGender, ncol = 1)
```

External activities
-------------------

```{r activities}
ggplot(research, aes(x = ExternalActivities)) +
    geom_bar() +
    coord_flip()
```

Facilities
----------

```{r facilities}
research %>%
    select(starts_with("Facilities_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Facilities_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By institute

```{r facilities-institute}
research %>%
    select(starts_with("Facilities_"), Institute) %>%
    group_by(Institute) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Institute) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Facilities_")) %>%
    group_by(Type, GroupSize, Institute) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Institute)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Institute, ncol = 1)
```

Financial support
-----------------

```{r finsupport}
research %>%
    select(starts_with("FinancialSupport_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "FinancialSupport_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By institute

```{r finsupport-institute}
research %>%
    select(starts_with("FinancialSupport_"), Institute) %>%
    group_by(Institute) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Institute) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "FinancialSupport_")) %>%
    group_by(Type, GroupSize, Institute) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Institute)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Institute, ncol = 1)
```

Other support
-------------

```{r othersupport}
research %>%
    select(starts_with("OtherSupport_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "OtherSupport_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By institute

```{r othersupport-institute}
research %>%
    select(starts_with("OtherSupport_"), Institute) %>%
    group_by(Institute) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Institute) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "OtherSupport_")) %>%
    group_by(Type, GroupSize, Institute) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Institute)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Institute, ncol = 1)
```

### By gender

```{r othersupport-gender}
research %>%
    select(starts_with("OtherSupport_"), Gender) %>%
    group_by(Gender) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Gender) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "OtherSupport_")) %>%
    group_by(Type, GroupSize, Gender) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Gender)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Gender, ncol = 1)
```

### By primary gender

```{r othersupport-primarygender}
research %>%
    select(starts_with("OtherSupport_"), PrimaryGender) %>%
    filter(PrimaryGender %in% c("Male", "Female")) %>%
    group_by(PrimaryGender) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -PrimaryGender) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "OtherSupport_")) %>%
    group_by(Type, GroupSize, PrimaryGender) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = PrimaryGender)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ PrimaryGender, ncol = 1)
```

Involvement
===========

COMBINE
-------

```{r combine}
responses %>%
    select(starts_with("COMBINE_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "COMBINE_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

ABACBS
------

```{r abacbs}
responses %>%
    select(starts_with("ABACBS_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "ABACBS_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

ABACBS Membership
-----------------

```{r membership}
responses %>%
    select(starts_with("Membership_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Membership_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By membership

```{r membership-membership}
research %>%
    select(starts_with("Membership_"), ABACBS_Member) %>%
    group_by(ABACBS_Member) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -ABACBS_Member) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Membership_")) %>%
    group_by(Type, GroupSize, ABACBS_Member) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = ABACBS_Member)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ ABACBS_Member, ncol = 1)
```

### By residency

```{r membership-residency}
research %>%
    select(starts_with("Membership_"), Residency) %>%
    group_by(Residency) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Residency) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Membership_")) %>%
    group_by(Type, GroupSize, Residency) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Residency)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Residency, ncol = 1)
```

### By gender

```{r membership-gender}
research %>%
    select(starts_with("Membership_"), Gender) %>%
    group_by(Gender) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Gender) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Membership_")) %>%
    group_by(Type, GroupSize, Gender) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Gender)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Gender, ncol = 1)
```

COMBINE Events
--------------

```{r events}
responses %>%
    select(starts_with("Events_")) %>%
    mutate(GroupSize = n()) %>%
    gather(key = "Type", value = "Has", -GroupSize) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Events_")) %>%
    group_by(Type, GroupSize) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = Type)) %>%
    ggplot(aes(x = Type, y = Prop)) +
    geom_col() +
    coord_flip()
```

### By membership

```{r events-membership}
research %>%
    select(starts_with("Events_"), ABACBS_Member) %>%
    group_by(ABACBS_Member) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -ABACBS_Member) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Events_")) %>%
    group_by(Type, GroupSize, ABACBS_Member) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = ABACBS_Member)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ ABACBS_Member, ncol = 1)
```

### By residency

```{r events-residency}
research %>%
    select(starts_with("Events_"), Residency) %>%
    group_by(Residency) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Residency) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Events_")) %>%
    group_by(Type, GroupSize, Residency) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Residency)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Residency, ncol = 1)
```

### By gender

```{r events-gender}
research %>%
    select(starts_with("Events_"), Gender) %>%
    group_by(Gender) %>%
    mutate(GroupSize = n()) %>%
    ungroup() %>%
    gather(key = "Type", value = "Has", -GroupSize, -Gender) %>%
    filter(Has == TRUE) %>%
    mutate(Type = str_remove(Type, "Events_")) %>%
    group_by(Type, GroupSize, Gender) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupSize) %>%
    arrange(Prop) %>%
    mutate(Type = factor(Type, levels = unique(Type))) %>%
    ggplot(aes(x = Type, y = Prop, fill = Gender)) +
    geom_col() +
    coord_flip() +
    facet_wrap(~ Gender, ncol = 1)
```

Session info
============

```{r session-info, cache = FALSE}
devtools::session_info()
```

```{r cleanup-docs, cache = FALSE}
doc.files <- c(list.files(pattern = "pdf"),
               list.files(pattern = "html"),
               list.files(pattern = "docx"))

for (file in doc.files) {
    file.rename(file, file.path("../docs", file))
}
```
