==> analysis/04_report.Rmd <==
---
title: "COMBINE National Student Survey 2018"
author: "COMBINE Australia"
date: '`r Sys.Date()`'
output:
  html_document:
    number_sections: yes
    theme: yeti
    toc: yes
    toc_float: yes
    df_print: paged
---

Code version: `r system("git log -1 --format=oneline | cut -d' ' -f1", intern = TRUE)`

```{r knitr, include = FALSE}
DOCNAME = "report"
knitr::opts_chunk$set(autodep        = TRUE,
                      cache          = FALSE,
                      cache.path     = paste0("cache/", DOCNAME, "/"),
                      cache.comments = TRUE,
                      echo           = FALSE,
                      error          = FALSE,
                      fig.align      = "center",
                      fig.path       = paste0("figures/", DOCNAME, "/"),
                      fig.width      = 10,
                      fig.height     = 8,
                      message        = FALSE,
                      warning        = FALSE)
```

```{r libaries, cache = FALSE}
library("here")
library("tidyverse")
library("cowplot")
library("ggbeeswarm")
```

```{r source, cache = FALSE}
source(here("R/manipulation.R"))
```

```{r theme, cache = FALSE}
combine_blue <- "#4a86e8"
gender_cols <- c("#4daf4a", "#ff7f00")

theme_set(
    theme_minimal(base_size = 14) +
    theme(plot.title = element_text(hjust = 0.5),
          strip.text = element_text(size = 16, colour = "white"),
          strip.background = element_rect(fill = combine_blue,
                                          colour = combine_blue))
)
```

Introduction
============

```{r load}
col_types <- cols(
    Timestamp                     = col_datetime(format = ""),
    Gender                        = col_character(),
    Age                           = col_integer(),
    IndigenousStatus              = col_character(),
    Location                      = col_character(),
    CourseLevel                   = col_character(),
    CourseYear                    = col_character(),
    Residency                     = col_character(),
    TimeCommitment                = col_character(),
    Background                    = col_character(),
    BackgroundGrouped             = col_character(),
    Competencies_Genetics         = col_integer(),
    Competencies_Biochemistry     = col_integer(),
    Competencies_Probability      = col_integer(),
    Competencies_Testing          = col_integer(),
    Competencies_LinearAlgebra    = col_integer(),
    Competencies_Algorithms       = col_integer(),
    Competencies_Programming      = col_integer(),
    Competencies_SoftwareDev      = col_integer(),
    Finances                      = col_character(),
    Finances_Government           = col_logical(),
    Finances_University           = col_logical(),
    Finances_TopUp                = col_logical(),
    Finances_PartTime             = col_logical(),
    Finances_FullTime             = col_logical(),
    Finances_Parental             = col_logical(),
    Finances_Savings              = col_logical(),
    Finances_HasOther             = col_logical(),
    Finances_Other                = col_character(),
    CourseworkComponent           = col_character(),
    CourseworkCurrent             = col_character(),
    Classes                       = col_double(),
    Classes_Computational         = col_integer(),
    Classes_Statistical           = col_integer(),
    Classes_Biological            = col_integer(),
    CourseworkTime                = col_double(),
    Assignments                   = col_integer(),
    Exams                         = col_integer(),
    Cohort                        = col_character(),
    Support                       = col_character(),
    CourseworkDetails             = col_character(),
    ResearchComponent             = col_character(),
    ResearchCurrent               = col_character(),
    ResearchPercentage            = col_integer(),
    ResearchTime                  = col_integer(),
    ResearchFocus                 = col_character(),
    Focus_GenomicVariants         = col_logical(),
    Focus_GeneExpression          = col_logical(),
    Focus_Metagenomics            = col_logical(),
    Focus_Proteomics              = col_logical(),
    Focus_DataVis                 = col_logical(),
    Focus_Methods                 = col_logical(),
    Focus_Cancer                  = col_logical(),
    Focus_Analysis                = col_logical(),
    Focus_CellBiology             = col_logical(),
    Focus_PlantBiology            = col_logical(),
    Focus_RareDisease             = col_logical(),
    Focus_Metabolomics            = col_logical(),
    Focus_MicrobialBiology        = col_logical(),
    Focus_HumanBiology            = col_logical(),
    Focus_ModelOrganisims         = col_logical(),
    Focus_NonModelOrganisims      = col_logical(),
    Focus_HasOther                = col_logical(),
    Focus_Other                   = col_character(),
    Institute                     = col_character(),
    Computational                 = col_integer(),
    Technologies                  = col_character(),
    Technologies_RNASeq           = col_logical(),
    Technologies_WholeGenome      = col_logical(),
    Technologies_Exome            = col_logical(),
    Technologies_ShortRead        = col_logical(),
    Technologies_LongRead         = col_logical(),
    Technologies_MassSpec         = col_logical(),
    Technologies_Array            = col_logical(),
    Technologies_FlowCytometry    = col_logical(),
    Technologies_HasOther         = col_logical(),
    Technologes_Other             = col_character(),
    ProgrammingLanguages          = col_character(),
    Programming_R                 = col_logical(),
    Programming_Python            = col_logical(),
    Programming_Shell             = col_logical(),
    Programming_Git               = col_logical(),
    Programming_Perl              = col_logical(),
    Programming_CCpp              = col_logical(),
    Programming_MATLAB            = col_logical(),
    Programming_HasOther          = col_logical(),
    Programming_Other             = col_character(),
    Collaboration                 = col_character(),
    Collaboration_YourGroup       = col_logical(),
    Collaboration_OtherGroups     = col_logical(),
    Collaboration_ExternalGroups  = col_logical(),
    Collaboration_OtherStudents   = col_logical(),
    ResearchGroup                 = col_character(),
    OtherStudents                 = col_character(),
    ResearchDetails               = col_character(),
    SupervisorNumber              = col_integer(),
    PrimaryPercentage             = col_double(),
    PrimaryGender                 = col_character(),
    SecondaryGender               = col_character(),
    FormalMeetings                = col_character(),
    InformalMeetings              = col_character(),
    Supervisor_Communication      = col_integer(),
    Supervisor_Inclusiveness      = col_integer(),
    Supervisor_Empathy            = col_integer(),
    Supervisor_SubjectMatter      = col_integer(),
    Supervisor_Availability       = col_integer(),
    Supervisor_Project            = col_integer(),
    ExternalActivities            = col_character(),
    FacilityAccess                = col_character(),
    Facilities_Desk               = col_logical(),
    Facilities_Computer           = col_logical(),
    Facilities_Server             = col_logical(),
    Facilities_Lab                = col_logical(),
    Facilities_MeetingRooms       = col_logical(),
    Facilities_BreakRooms         = col_logical(),
    FinancialSupport              = col_character(),
    FinancialSupport_TravelApp    = col_logical(),
    FinancialSupport_TravelNoApp  = col_logical(),
    FinancialSupport_TopUp        = col_logical(),
    FinancialSupport_HasOther     = col_logical(),
    FinancialSupport_Other        = col_character(),
    OtherSupport                  = col_character(),
    OtherSupport_Group            = col_logical(),
    OtherSupport_Institute        = col_logical(),
    OtherSupport_Students         = col_logical(),
    OtherSupport_People           = col_logical(),
    OtherSupport_None             = col_logical(),
    SupervisorDetails             = col_character(),
    FamiliarityCOMBINE            = col_character(),
    COMBINE_Survey                = col_logical(),
    COMBINE_SocialMedia           = col_logical(),
    COMBINE_Newsletter            = col_logical(),
    COMBINE_Workshop              = col_logical(),
    COMBINE_Seminar               = col_logical(),
    COMBINE_Symposium             = col_logical(),
    FamiliarityABACBS             = col_character(),
    ABACBS_Member                 = col_logical(),
    ABACBS_SocialMedia            = col_logical(),
    ABACBS_Newsletter             = col_logical(),
    ABACBS_Conference             = col_logical(),
    ABACBS_Event                  = col_logical(),
    ABACBS_None                   = col_logical(),
    MembershipABACBS              = col_character(),
    Membership_National           = col_logical(),
    Membership_EventDiscount      = col_logical(),
    Membership_ConferenceDiscount = col_logical(),
    Membership_Travel             = col_logical(),
    Membership_Prizes             = col_logical(),
    Membership_CV                 = col_logical(),
    Membership_Direction          = col_logical(),
    Membership_NoBenefits         = col_logical(),
    EventsCOMBINE                 = col_character(),
    Events_Programming            = col_logical(),
    Events_Bioinformatics         = col_logical(),
    Events_Skills                 = col_logical(),
    Events_Statistics             = col_logical(),
    Events_Lab                    = col_logical(),
    Events_OneOff                 = col_logical(),
    Events_InvitedSeminar         = col_logical(),
    Events_StudentSeminar         = col_logical(),
    Events_Social                 = col_logical(),
    Events_Symposium              = col_logical(),
    Events_NotInterested          = col_logical(),
    Events_HasOther               = col_logical(),
    Events_Other                  = col_character(),
    WhatCanWeDo                   = col_character()
)

responses <- read_tsv(here("data/responses_clean.tsv"),
                      col_types = col_types)
```

This year, for the first time, COMBINE ran a national survey of students.
The goal of the survey was to learn more about the Australian bioinformatics and
computational biology community, specifically:

- Who is studying bioinformatics and computational biology in Australia
- Their academic backgrounds and skills
- The areas of research and technologies they are working with
- The support they receive from their supervisors, institutes and universities
- What issues they encounter during their studies
- What COMBINE can do to support the student community

The survey opened in April and closed in June and we received over 80 responses.
Thank you to everyone who completed the survey and congratulations to those that
won a gift voucher in our prize draw.

Included here are some of the interesting results from the survey. When
interpreting what is shown here it is important to remember that this is a
fairly small and selective sample. As a result the data is more a reflection of 
the COMBINE community than the wider student community, however, we still
believe this has been a useful and informative exercise. We also believe that
this is a valuable opportunity for students to see what others' experiences are 
like and contrast them with their own.

Who completed the survey?
=========================

We received `r nrow(responses)` responses to the survey, after removing a couple 
that were duplicated or incomplete. The majority of these were from full-time
PhD students, however there were some other course types and a few part-time 
responses. The sample was also biased towards female and domestic students.

```{r demographics}
p1 <- ggplot(responses, aes(x = Gender, y = ..prop.., group = 1)) + 
    geom_bar(fill = combine_blue) + 
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Gender") +
    theme(axis.title = element_blank())

p2 <- ggplot(responses, aes(x = Residency, y = ..prop.., group = 1)) + 
    geom_bar(fill = combine_blue) + 
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Residency") +
    theme(axis.title = element_blank())

p3 <- ggplot(responses, aes(x = Location, y = ..prop.., group = 1)) + 
    geom_bar(fill = combine_blue) + 
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Location") +
    theme(axis.title = element_blank())

p4 <- responses %>%
    mutate(CourseLevel = factor(CourseLevel)) %>%
    mutate(CourseLevel = fct_relabel(CourseLevel,
                                     ~ str_replace(.x, " ", "\n"))) %>%
    ggplot(aes(x = CourseLevel, y = ..prop.., group = 1)) + 
    geom_bar(fill = combine_blue) + 
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Course Type") +
    theme(axis.title = element_blank())

plot_grid(p1, p2, p3, p4, ncol = 2)
```

Perhaps the most obvious and unexpected demographic is in the distribution of
student locations, particularly the high numbers from Victoria and the low
numbers from NSW. This is likely to be at least partly a result of COMBINE's
history as we were based in Melbourne for a long time and have had many
Victorian members of the executive. We are hoping to address this in the second
half of the year by running more events in Sydney. If you are based in NSW and
would like to get involved we would love to hear from you, particularly as the
COMBINE Symposium will be held in Sydney in 2019.

We also asked students about their academic backgrounds. The majority were
coming into bioinformatics from biology, but male students were more 
likely to come from a computer science or software engineering background.

```{r background}
responses %>%
    select(Gender, BackgroundGrouped) %>%
    mutate(BackgroundGrouped = factor(BackgroundGrouped)) %>%
    mutate(BackgroundGrouped = fct_relabel(BackgroundGrouped,
                                           ~ str_replace(.x, "/", "\n")),
           BackgroundGrouped = fct_relevel(BackgroundGrouped, "Other",
                                           after = Inf)) %>%
    group_by(Gender) %>%
    mutate(GroupCount = n()) %>%
    group_by(Gender, BackgroundGrouped, GroupCount) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / GroupCount) %>%
    ggplot(aes(x = BackgroundGrouped, y = Prop, fill = Gender)) + 
    geom_col(position = "dodge") + 
    #facet_wrap(~ Gender, ncol = 1) +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_manual(values = gender_cols) +
    ggtitle("Background") +
    theme(axis.title = element_blank())
```

Other backgrounds included pharmacy, biotechnology, psychology and various types
of engineering. It is unsurprising to see many backgrounds as bioinformatics has
traditionally drawn from many fields, but it will be interesting to see if this
changes as more computational subjects are included in undergraduate degrees.

Bioinformatics and computational biology can require a diverse skill set
including aspects of biology, computer science and statistics. We asked students
to rate how competent they felt at a variety of skills on a scale from 0
(lowest) to 5 (highest).

```{r competencies, fig.height = 12}
p1 <- responses %>%
    select(Gender, starts_with("Competencies")) %>%
    gather(key = "Competency", value = "Rating", -Gender) %>%
    mutate(Competency = str_remove(Competency, "Competencies_")) %>%
    group_by(Gender, Competency, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Competency, y = Prop, fill = fct_rev(factor(Rating)))) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Blues", name = "Rating", direction = -1) +
    ggtitle("Competencies by gender") +
    facet_wrap(~ Gender, ncol = 1) +
    theme(axis.text.x = element_blank(),
          axis.title = element_blank(),
          legend.position = "none")

p2 <- responses %>%
    select(Residency, starts_with("Competencies")) %>%
    gather(key = "Competency", value = "Rating", -Residency) %>%
    mutate(Competency = str_remove(Competency, "Competencies_")) %>%
    group_by(Residency, Competency, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ungroup() %>%
    mutate(Competency = fct_recode(Competency,
                                   `Linear\nAlgebra` = "LinearAlgebra",
                                   `Software\nDevelopment` = "SoftwareDev")) %>%
    ggplot(aes(x = Competency, y = Prop, fill = fct_rev(factor(Rating)))) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Blues", name = "Rating", direction = -1) +
    guides(fill = guide_legend(nrow = 1)) +
    ggtitle("Competencies by residency") +
    facet_wrap(~ Residency, ncol = 1) +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
          axis.title = element_blank(),
          legend.position = "bottom")

plot_grid(p1, p2, ncol = 1, rel_heights = c(0.45, 0.55))
```

Overall we saw quite a spread of ratings for all skills. Female students were
more confident in their genetics skills while male students were more
comfortable with programming and software development. International students 
rated themselves more highly at many of the technical skills including
computational algorithms, linear algebra, programming and software development.

Money can be a source of stress for many students and we were interested in
how bioinformatics students supported themselves. In general scholarships
(university and government) and part-time work were the biggest sources of
financial support, but unsurprisingly this is different when factors such as
course type or residency are considered.

```{r finances, fig.width = 8, fig.height = 9}
p1 <- responses %>%
    summarise_category("Finances_") %>%
    mutate(Key = fct_recode(Key,
                            `Full-time\nwork` = "FullTime",
                            `Government\nscholarship` = "Government",
                            Other = "HasOther",
                            `Parental\nsupport` = "Parental",
                            `Part-time\nwork` = "PartTime",
                            `Top-up\nscholarship` = "TopUp",
                            `University\nscholarship` = "University")) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("Financial support") +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
          legend.position = "none")

p2 <- responses %>%
    summarise_category_grouped("Finances_", Gender) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By gender") +
    facet_wrap(~ Gender, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

p3 <- responses %>%
    summarise_category_grouped("Finances_", Residency) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By residency") +
    facet_wrap(~ Residency, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

p4 <- responses %>%
    summarise_category_grouped("Finances_", CourseLevel) %>%
    filter(CourseLevel %in% c("Masters (Coursework)", "PhD")) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By course") +
    facet_wrap(~ CourseLevel, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

plot_grid(p1, p2, p3, p4, ncol = 2, nrow = 2)
```

Domestic students are much more likely to be supported by government
scholarships or part-time work while international students typically rely on 
university funding. Master's students rely more on part-time work and parental 
support, but worryingly half also responded that they supported themselves using 
their savings. We neglected to included government welfare as an option in the 
survey but this was recorded in other responses, along with scholarships from
other funding bodies.

What are students studying?
===========================

Of the `r nrow(responses)` responses,
`r sum(responses$CourseworkCurrent == "Yes")` were currently completing
coursework classes. The majority of these were Master by coursework students
but they covered all levels from Bachelor to PhD.

```{r coursework}
coursework <- filter(responses, CourseworkCurrent == "Yes")

coursework %>%
    select(CourseLevel, Classes, CourseworkTime, Assignments, Exams) %>%
    gather(key = "Key", value = "Value", -CourseLevel, -Classes) %>%
    mutate(Classes = factor(Classes)) %>%
    mutate(Key = factor(Key,
                        levels = c("CourseworkTime", "Assignments", "Exams"),
                        labels = c("Class hours", "Assignments", "Exams"))) %>%
    ggplot(aes(x = Classes, y = Value, colour = CourseLevel)) +
    geom_beeswarm(size = 5, alpha = 0.8, cex = 3) +
    scale_colour_brewer(palette = "Set1", name = "Course type") +
    coord_flip() +
    facet_wrap(~ Key, nrow = 1, scales = "free_x") +
    ggtitle("Coursework details") +
    xlab("Number of classes") +
    theme(legend.position = "bottom",
          axis.title.x = element_blank(),
          panel.border = element_rect(fill = NA, colour = combine_blue))
```

Coursework students were enrolled in one to four classes in the first semester
of 2018 with between one and 30 hours of classes a week. There was considerable 
variation in how classes were assessed with students having up to 16 assignments 
in a semester and up to six exams. 


We asked students what percentage of their current classes contained biological,
statistical or computational content. One class may fit into two categories,
for example a statistics subject might teach statistical theory but also how
to implement it computationally, meaning that the total could come to more than
100 percent.

```{r class-types}
coursework %>%
    select(Classes_Computational, Classes_Statistical, Classes_Biological) %>%
    filter(!is.na(Classes_Computational)) %>%
    arrange(Classes_Statistical, Classes_Computational, Classes_Biological) %>%
    mutate(Person = row_number()) %>%
    gather(key = Type, value = Percent, -Person) %>%
    mutate(Type = str_remove(Type, "Classes_")) %>%
    ggplot(aes(x = factor(Person), y = Percent, fill = Type)) +
    geom_col() +
    scale_y_continuous(labels = function(x) {paste0(x, "%")}) +
    scale_fill_brewer(palette = "Set1", name = "Class type") +
    ggtitle("Types of classes") +
    xlab("Students") +
    ylab("Percentage of classes") +
    theme(legend.position = "bottom",
          axis.text.x = element_blank())
```

We found that classes are fairly evenly spread through the three disciplines, 
although there are less interdisciplinary subjects than we expected. This may 
represent a truth about the fragmented nature of universities or it could be 
related to how these questions were worded.

Detailed comments provided by students give more insight into the coursework
experience, particularly some of the negative aspects. One student commented
that their Master's course has "No support. No prerequisites" and was "Badly
designed", while others focused on assessment: "Too many assignments. Interrupts
meaningful learning". Other students felt they were not being adequately
prepared for the field: "... the coursework is geared towards experimental
projects with no bioinformatics or statistics unfortunately". One detailed
response touches on many of the difficulties for coursework students, 
particularly those from biology backgrounds:

> "The pre-requisite mathematics knowledge for the biology stream are far lower
> than what is expected from all of the probability & statistics courses. There 
> are no bridging courses, and no support for students who are having difficulty 
> in these courses. Most biology stream students who have taken these courses 
> have struggled with the content, even when they put 90% of their time into 
> studying for these courses at the expense of others. Given the grades
> necessary to enter a PhD program with a scholarship, it feels like being set
> up to fail."

This comment raises some considerable concerns and suggests that universities
continue to grapple with preparing students for modern, data-focused biology.
We expect that universities will continue to develop and improve  
interdisciplinary courses and integrate more statistical and computational
skills into undergraduate biology courses, hopefully leading to less of these
experiences in the future. 

What are students researching?
==============================

```{r research}
research <- filter(responses, ResearchCurrent == "Yes")
```

The majority of responses were from students who are currently completing a
research project (`r sum(responses$ResearchCurrent == "Yes")` of
`r nrow(responses)`). These projects cover a range of research focuses. As we
would expect of bioinformatics students the majority are involved in data
analysis, with methods development and data visualisation also common. Of the
research topics, gene expression and genomic variants were most common but there
was quite a range of other focuses in student research.

Short-read sequencing technologies including RNA-seq and whole genome
sequencing were the most common sources of data. Other technologies not given
as options include ChIP-seq and ATAC-seq.

```{r focus, fig.height = 10}
p1 <- research %>%
    summarise_category("Focus_") %>%
    mutate(Key = fct_recode(Key,
                            `Data analysis` = "Analysis",
                            `Methods\ndevelopment` = "Methods",
                            `Gene expression` = "GeneExpression",
                            `Data\nvisualisation` = "DataVis",
                            `Genomic\nvariants` = "GenomicVariants",
                            `Human biology` = "HumanBiology",
                            `Microbial biology` = "MicrobialBiology",
                            `Cell biology` = "CellBiology",
                            `Model\norganisms` = "ModelOrganisims",
                            `Plant biology` = "PlantBiology",
                            `Non-model\norganisms` = "NonModelOrganisims",
                            `Rare disease` = "RareDisease",
                            `Other` = "HasOther")) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Student research focuses") +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

p2 <- research %>%
    summarise_category("Technologies_") %>%
    mutate(Key = fct_recode(Key,
                            `Short-read\nsequencing` = "ShortRead",
                            `RNA-seq` = "RNASeq",
                            `Whole genome\nsequencing` = "WholeGenome",
                            `Mass\nspectrometry` = "MassSpec",
                            `Long read\nsequencing` = "LongRead",
                            `Array-based\ntechnologies` = "Array",
                            `Exome\nsequencing` = "Exome",
                            `Flow cytometry` = "FlowCytometry",
                            `Other` = "HasOther")) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Technologies used by students") +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

p3 <- research %>%
    summarise_category("Programming_") %>%
    mutate(Key = fct_recode(Key,
                            `C/C++` = "CCpp",
                            `Other` = "HasOther")) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Programming languages and tools used by students") +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

plot_grid(p1, p2, p3, ncol = 1)
```

A core set of four programming languages were used by students in their
research (R, Shell, Python and Git), with 80 percent of students using R. Other
tools are used by some students but these four are clearly the current dominant
programming tools.

While many students use the same sets of computational tools there is 
considerable variation in the percentage of their projects that are 
computational.

```{r computational-time}
p1 <- ggplot(research, aes(x = Computational)) +
    geom_density(colour = combine_blue, fill = combine_blue) +
    scale_x_continuous(labels = function(x) {paste0(x, "%")}) +
    facet_wrap(~ Gender) +
    ylab("Density") +
    ggtitle("Percentage of projects that is computational") +
    theme(axis.title.x = element_blank(),
          axis.text.x = element_blank())

p2 <- ggplot(research, aes(x = Computational)) +
    geom_density(colour = combine_blue, fill = combine_blue) +
    scale_x_continuous(labels = function(x) {paste0(x, "%")}) +
    facet_wrap(~ Residency) +
    ylab("Density") +
    theme(axis.title.x = element_blank())

plot_grid(p1, p2, ncol = 1)
```

Male students and international students were more likely to have purely
computational projects while female and domestic students showed a greater
spread. This may be related to students' academic background with female and 
domestic students more likely to have a biology background and the related lab 
skills.

We also saw differences between students as to who they were collaborating with
for their research.

```{r collaboration}
research %>%
    summarise_category_grouped("Collaboration_", Gender, has_other = FALSE) %>%
    mutate(Key = fct_recode(Key,
                            `People in\nyour group` = "YourGroup",
                            `Other groups\nat your institute` = "OtherGroups",
                            `Groups at\nother institutes` = "ExternalGroups",
                            `Other students` = "OtherStudents")) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    facet_wrap(~ Gender) +
    ggtitle("Who do students collaborate with?") +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))
```

About twice as many male students were collaborating with other groups at their 
university or institute while more female students collaborated with other
students. 

Who are their supervisors?
==========================

Supervisors can be just as crucial to the success of a research project as the
students themselves. The majority of research students had either two or three
supervisors with primary supervisors being responsible for two-thirds of
supervision on average.

```{r supervisor-gender}
research %>%
    select(Gender, PrimaryGender, SecondaryGender) %>%
    rename(Primary = "PrimaryGender", Secondary = "SecondaryGender") %>%
    group_by(Gender) %>%
    mutate(TotalCount = n()) %>%
    group_by(Gender, TotalCount, Primary, Secondary) %>%
    summarise(Count = n()) %>%
    ungroup() %>%
    mutate(Prop = Count / TotalCount) %>%
    filter(Primary %in% c("Female", "Male"),
           Secondary %in% c("Female", "Male")) %>%
    ggplot(aes(x = Gender, y = Prop, fill = Gender)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_manual(values = gender_cols, name = "Student gender") +
    facet_grid(Secondary ~ Primary,
               labeller = partial(label_both, sep = "\n"), margins = TRUE) +
    ggtitle("Supervisor genders") +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "bottom")
```

The majority of primary supervisors (77 percent) were male. Female students were
slightly more likely to have female primary supervisors but were also more
likely to have male secondary supervisors. It is likely that the prevalence of 
male supervisors is a result of the gender imbalance at higher levels of
academia rather than any preference from students.

We asked students to rate their supervisors on a range of qualities that can
lead to good student-supervisor relationships and successful projects.

```{r qualities}
research %>%
    select(PrimaryGender, starts_with("Supervisor_")) %>%
    filter(PrimaryGender %in% c("Male", "Female")) %>%
    mutate(PrimaryGender = paste(PrimaryGender, "supervisors")) %>%
    gather(key = "Quality", value = "Rating", -PrimaryGender) %>%
    mutate(Quality = str_remove(Quality, "Supervisor_")) %>%
    group_by(PrimaryGender, Quality, Rating) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ggplot(aes(x = Quality, y = Prop, fill = fct_rev(factor(Rating)))) +
    geom_col() +
    scale_x_discrete(labels = c("Availability", "Communication", "Empathy",
                                "Inclusiveness", "Interest in\nproject",
                                "Subject matter\nexpertise")) +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Blues", name = "Rating", direction = -1) +
    ggtitle("Supervisor qualities") +
    facet_wrap(~ PrimaryGender, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
          legend.position = "bottom")
```

Female primary supervisors were rated more highly on most qualities, with the
exception of subject matter expertise. It is important to note that these
ratings are subjective and may reflect the biases of students. Overall these
ratings are high, suggesting that most supervisors are generally doing a good 
job. However some students had issues, particularly with supervisors without
computational backgrounds. For example, one student mentioned their supervisor
had a "...lack of expertise in the field combined with lack of collaboration
with other groups/researchers, and even not allowing me to consult other experts
in the field ...", and another "Supervisor has no computational background so 
getting guidance on this area is challenging".

Other students had mixed experiences with their supervisors, highlighting why it
can be important to have more than one:

> "My primary supervisor is truely wonderful. She's super supportive, engaged,
> and even though she is busy, always makes time for me. Although I have a 
> secondary supervisor, I have had maybe 2 meetings with her my whole PhD (2.5 
> years in) and she hasn't attended any of my official events (confirmation, 
> review, presentations etc.)."

What support do students receive?
=================================

We asked coursework students how supported they felt by their university and
student cohort.

```{r coursework-support}
coursework %>%
    rename(Support_Cohort = Cohort, Support_Uni = Support) %>%
    mutate(Support_Cohort = if_else(Support_Cohort == "Yes", TRUE, FALSE),
           Support_Uni = if_else(Support_Uni == "Yes", TRUE, FALSE)) %>%
    summarise_category_grouped("Support_", Gender, has_other = FALSE) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_x_discrete(labels = c("Part of a\nstudent cohort",
                                "Adequate university\nsupport")) +
    ggtitle("Support for coursework students") +
    ylab('Percentage of students who answered "Yes"') +
    facet_wrap(~ Gender) +
    theme(axis.title.x = element_blank())
```

Significantly more female students answered yes to both questions. While all 
female students felt that they were part of a student cohort this was true for
less than 75 percent of male students.

How supported research students feel can depend more on their supervisor and
research group than their university.

```{r group-support}
p1 <- ggplot(research, aes(x = ResearchGroup, y = ..prop.., group = 1)) +
    geom_bar(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Do students feel part of their research group?") +
    theme(axis.title = element_blank())

p2 <- ggplot(research, aes(x = OtherStudents, y = ..prop.., group = 1)) +
    geom_bar(fill = combine_blue) +
    scale_x_discrete(labels = c("No, I am the only\nstudent in my group",
                                "Yes, but they don't\ndo computational work",
                                "Yes, there are other\ncomputational students")) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Are there other students in your research group?") +
    theme(axis.title = element_blank())

plot_grid(p1, p2, ncol = 1)
```

Most research students felt like they were part of their research groups, 
one student commenting "I have a fantastic relationship with my supervisor and 
fellow lab-members. I couldn't be happier with my research environment".
However, for the minority of students that felt differently, this seemed to be
related to the cohesiveness of their groups with students explaining that their 
research group was "... not tight or supportive enough. Everyone working 
independently with basically no interaction ..." and  "It's very exclusive, and 
there is close to zero research interaction in the group (everyone works 
individually)". These feeling could also be related to the 15 percent of
students that are alone in their research groups. 

We also asked students more broadly about their sources of support outside their
supervisors. Overall 73 percent of students felt supported by their research
group, 65 percent by their institute and around 55 percent by other people or
students in their lives. These values are lower than we would hope to see and
ideally every student would benefit from all these sources of support. Most
worryingly 8 percent of students said they had no source of support outside
their supervisor.

```{r research-support, fig.width = 8, fig.height = 12}
p1 <- research %>%
    summarise_category("OtherSupport_", has_other = FALSE) %>%
    mutate(Key = fct_recode(Key,
                            `Research group` = "Group",
                            `Other people` = "People",
                            `Other students` = "Students")) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("Sources of support") +
    theme(axis.title = element_blank(),
          axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
          legend.position = "none")

p2 <- research %>%
    summarise_category_grouped("OtherSupport_", Gender, has_other = FALSE) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By gender") +
    facet_wrap(~ Gender, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

p3 <- responses %>%
    summarise_category_grouped("OtherSupport_", Residency,
                               has_other = FALSE) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By residency") +
    facet_wrap(~ Residency, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

p4 <- responses %>%
    summarise_category_grouped("OtherSupport_", Institute,
                               has_other = FALSE) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By institute") +
    facet_wrap(~ Institute, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

p5 <- responses %>%
    summarise_category_grouped("OtherSupport_", PrimaryGender,
                               has_other = FALSE) %>%
    filter(PrimaryGender %in% c("Male", "Female")) %>%
    ggplot(aes(x = Key, y = Prop, fill = Key)) +
    geom_col() +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_brewer(palette = "Set1") +
    ggtitle("By supervisor gender") +
    facet_wrap(~ PrimaryGender, ncol = 1) +
    theme(axis.title = element_blank(),
          axis.text.x = element_blank(),
          legend.position = "none")

pp <- plot_grid(p2, p3, p4, p5, ncol = 2, nrow = 2)
plot_grid(p1, pp, ncol = 1, rel_heights = c(1, 2))
```

Female students generally felt more supported, as did domestic students and
those with female primary supervisors. Students associated with, or based at,
a research institute were more likely to feel supported by their research group
than those based purely at a university. 

One way that support can come from supervisors is through meetings with
students. We asked how often student have formal (scheduled meetings with the
research project as the main topic) and informal (getting coffee, asking 
questions, coming by a desk etc.) interaction with their supervisors. 

```{r meetings}
research %>%
    select(Gender, FormalMeetings, InformalMeetings) %>%
    gather(key = "Type", value = "Frequency", -Gender) %>%
    mutate(Type = str_remove(Type, "Meetings")) %>%
    mutate(Frequency = fct_relevel(Frequency,
                                   "Every day",
                                   "A few times a week",
                                   "Once a week",
                                   "Once a fortnight",
                                   "Once a month",
                                   "Once every six months",
                                   "Once a year",
                                   "I never see my supervisor(s)"),
           Frequency = fct_recode(
               Frequency,
               `A few times\na week` = "A few times a week",
               `Once a\nfortnight` = "Once a fortnight",
               `Once every\nsix months` = "Once every six months",
               `I never see my\nsupervisor(s)` =
                   "I never see my supervisor(s)")) %>%
    ggplot(aes(x = fct_rev(Frequency), y = ..prop..,
               fill = Gender, group = Gender)) +
    geom_bar(position = "dodge") +
    scale_y_continuous(labels = scales::percent_format()) +
    scale_fill_manual(values = gender_cols, name = "Student gender") +
    coord_flip() +
    facet_wrap(~ Type, nrow = 1) +
    ggtitle("Frequency of supervisor meetings") +
    theme(axis.title = element_blank(),
          legend.position = "bottom")
```

Most male students had formal meetings once a fortnight and informal
interactions a few times a week. Female students were more variable, meeting
both more and less frequently. Meeting frequency may depend on workload and
preference as this comment notes: "If I have a lot of trouble and questions I 
might see them everyday but then there are periods where I maybe see them once a 
month. I really like the freedom of that".

Access to appropriate facilities is also vital for students to successfully
complete a research project. We are pleased to see that the majority of students 
are provided with desks and computers as well as access to servers or
high-performance computing that can be essential for many bioinformatics
projects.

```{r institute-support}
p1 <- research %>%
    summarise_category_grouped("Facilities_", Institute, has_other = FALSE) %>%
    mutate(Key = fct_recode(Key,
                            `Meeting rooms` = "MeetingRooms",
                            `Break rooms` = "BreakRooms",
                            `Laboratory` = "Lab")) %>%
    ggplot(aes(x = fct_rev(Key), y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    facet_wrap(~ Institute, ncol = 1) +
    coord_flip() +
    ggtitle("Access to facilities") +
    theme(axis.title = element_blank())

p2 <- research %>%
    summarise_category_grouped("FinancialSupport_", Institute,
                               has_other = FALSE) %>%
    mutate(Key = fct_recode(Key,
                            `Travel funding\n(with application)` = "TravelApp",
                            `Top-up\nscholarships` = "TopUp",
                            `Travel funding\n(no application)` = "TravelNoApp",
                            `Other` = "HasOther")) %>%
    ggplot(aes(x = fct_rev(Key), y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_y_continuous(labels = scales::percent_format()) +
    facet_wrap(~ Institute, ncol = 1) +
    coord_flip() +
    ggtitle("Access to financial support") +
    theme(axis.title = element_blank())

plot_grid(p1, p2, nrow = 1)
```

Funding to attend conferences can also be important, particularly for PhD
students hoping to continue in academia and thereofore needing opportunities to 
present their work and meet people in the field. Most students had access to
travel grants with an application through their institute or university, but
those based at research institutes had greater access to top-up scholarships.

What can COMBINE do to help?
============================

The last section of the survey was designed to find out how COMBINE can help to
improve the student experience. We started by asking students what previous
experience they had had with COMBINE or ABACBS.

```{r involvement}
p1 <- responses %>%
    summarise_category("COMBINE_", has_other = FALSE) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_x_discrete(labels = c("Newsletter", "Survey", "Social media",
                                "Seminar", "Symposium", "Workshop")) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Previous involvement with COMBINE") +
    theme(axis.title = element_blank())

p2 <- responses %>%
    summarise_category("ABACBS_", has_other = FALSE) %>%
    ggplot(aes(x = Key, y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_x_discrete(labels = c("Member", "Newsletter", "Social media",
                                "None", "Conference", "Event")) +
    scale_y_continuous(labels = scales::percent_format()) +
    ggtitle("Previous involvement with ABACBS") +
    theme(axis.title = element_blank())

plot_grid(p1, p2, ncol = 1)
```

Students were most likely to connect with COMBINE and ABACBS via mailing lists
and social media and around a quarter had attended the COMBINE symposium or
ABACBS conference.

About half of students reported they were paid ABACBS members and we asked what
they thought the benefits of membership were.

```{r membership}
p1 <- responses %>%
    summarise_category_grouped("Membership_", ABACBS_Member,
                               has_other = FALSE) %>%
    mutate(ABACBS_Member = if_else(ABACBS_Member, "Member", "Non-member")) %>%
    ggplot(aes(x = fct_rev(Key), y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_x_discrete(labels = c("No benefits\nfor students",
                                "Say in the direction\nof the society",
                                "Looks good on a CV",
                                "Access to\nprizes and awards",
                                "Access to\ntravel bursaries",
                                "Member of a\nnational society",
                                "Conference discounts",
                                "Event discounts")) +
    scale_y_continuous(labels = scales::percent_format(), limits = c(0, 0.85)) +
    coord_flip() +
    facet_wrap(~ ABACBS_Member, nrow = 1) +
    ggtitle("Benefits of ABACBS membership") +
    theme(axis.title = element_blank())

p2 <- responses %>%
    summarise_category_grouped("Membership_", Residency,
                               has_other = FALSE) %>%
    ggplot(aes(x = fct_rev(Key), y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_x_discrete(labels = c("No benefits\nfor students",
                                "Say in the direction\nof the society",
                                "Looks good on a CV",
                                "Access to\nprizes and awards",
                                "Access to\ntravel bursaries",
                                "Member of a\nnational society",
                                "Conference discounts",
                                "Event discounts")) +
    scale_y_continuous(labels = scales::percent_format(), limits = c(0, 0.85)) +
    coord_flip() +
    facet_wrap(~ Residency, nrow = 1) +
    theme(axis.title = element_blank())

plot_grid(p1, p2, ncol = 1)
```

Discounts for events and conferences received the most positive responses as
demonstrated by these comments: "I think the most important thing is decrease
the registration fee of conferences" and "... provide more funding support to 
attend symposium". Access to prizes, awards and travel bursaries was
perceived to be more beneficial by non-members than members. International
students were generally less positive about membership benefits, with the
exception of it looking good on a CV.

When asked what future COMBINE events students would like to see, workshops
were the most popular option, along with the COMBINE Symposium. The general 
programming and introductory bioinformatics workshops we have run in the past 
have been successful, but clearly there is demand for more as evidenced by this 
comment: "I would love to see more beginner workshops, as I really benefitted 
from a workshop ran by COMBINE when I first started sequence analysis". 
Statistical workshops are not something we have done in the past but are
something to consider if we can find a format and content that works.

```{r combine-events}
responses %>%
    summarise_category("Events_", has_other = FALSE) %>%
    ggplot(aes(x = fct_rev(Key), y = Prop)) +
    geom_col(fill = combine_blue) +
    scale_x_discrete(labels = c("Not interested\nin events",
                                "Laboratory skills\nworkshops",
                                "One off seminars",
                                "Local social nights",
                                "Regular seminars\n(student speakers)",
                                "Regular seminars\n(invited speakers)",
                                "Professional skills\nworkshops",
                                "Programming workshops",
                                "COMBINE Symposium",
                                "Statistics workshops",
                                "Bioinformatics\nworkshops")) +
    scale_y_continuous(labels = scales::percent_format()) +
    coord_flip() +
    ggtitle("What possible COMBINE events would you like to attend?") +
    theme(axis.title = element_blank())
```

Local social nights and student seminar series are the most common types of 
events that COMBINE runs but these were of interest for less than 35 percent of
students that completed the survey. We believe these events are important for
building a student community but there is more we can do to make them more
attractive for more people. For example, one comment suggested that COMBINE
should "... hold monthly bioinformatics meet up in each city and the program 
includes allocating desks/tags/etc for different topics (eg. DNA seq, RNA seq, 
etc.) for asking questions, networking, etc.".

Other suggestions included mentoring programs, advocacy for student talks and
travel funding and a PhD award (stay tuned!). One comment asked us to provide 
"... a framework where students can connect with others for help and support
with aspects of their project, e.g. questions on programming (language specific
groups), troubleshooting advice, beta testers, etc.". Earlier this year we
opened up our Slack workspace for any students to join and we hope it can become 
exactly that kind of platform, as well as giving students more opportunity to
get involved with COMBINE.

On a final note, we really appreciated receiving comments like this, which we
think embodies exactly what COMBINE is trying to achieve:

> "COMBINE is the greatest. I value the student-driven focus and identity of
> COMBINE, as well as its role as an ABACBS sub-committee which enables cohesion 
> with the broader community whilst allowing students a space to develop 
> community leadership skills. Keep doing what you are doing!"

Session info
============

```{r session-info, cache = FALSE}
devtools::session_info()
```

```{r cleanup-docs, cache = FALSE}
doc.files <- c(list.files(pattern = "pdf"),
               list.files(pattern = "html"),
               list.files(pattern = "docx"))

for (file in doc.files) {
    file.rename(file, file.path("../docs", file))
}
```
